CSFD-121 code review: refactor abstract to static Util class for easy general usage
