+++ src/test/java/uk/gov/hmcts/ccd/v2/external/resource/CaseDataResourceTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.hateoas.Link;
 import uk.gov.hmcts.ccd.config.JacksonUtils;
 import uk.gov.hmcts.ccd.domain.model.std.CaseDataContent;
-import uk.gov.hmcts.ccd.v2.AbstractBaseResourceTest;
+import uk.gov.hmcts.ccd.util.PathFromUrlUtil;
 
 import java.util.Map;
 import java.util.Optional;
@@ -20,7 +20,7 @@ import static uk.gov.hmcts.ccd.domain.service.common.TestBuildersUtil.CaseDataBu
 import static uk.gov.hmcts.ccd.domain.service.common.TestBuildersUtil.CaseDataContentBuilder.newCaseDataContent;
 
 @DisplayName("CaseDataResource")
-class CaseDataResourceTest extends AbstractBaseResourceTest {
+class CaseDataResourceTest {
     private static final JsonNodeFactory JSON_NODE_FACTORY = new JsonNodeFactory(false);
     private static final ObjectMapper MAPPER = new ObjectMapper();
 
@@ -53,7 +53,7 @@ class CaseDataResourceTest extends AbstractBaseResourceTest {
         final CaseDataResource caseDataResource = new CaseDataResource(CASE_DATA_CONTENT, CASE_TYPE_ID, PAGE_ID);
 
         Optional<Link> self = caseDataResource.getLink("self");
-        assertThat(getActualPath(self.get().getHref()), equalTo(linkSelfForCaseData));
+        assertThat(PathFromUrlUtil.getActualPath(self.get().getHref()), equalTo(linkSelfForCaseData));
     }
 
 }
