+++ src/test/java/uk/gov/hmcts/ccd/v2/internal/resource/CaseHistoryViewResourceTest.java
@@ -11,7 +11,7 @@ import uk.gov.hmcts.ccd.domain.model.aggregated.CaseViewEvent;
 import uk.gov.hmcts.ccd.domain.model.aggregated.CaseViewField;
 import uk.gov.hmcts.ccd.domain.model.aggregated.CaseViewTab;
 import uk.gov.hmcts.ccd.domain.model.aggregated.CaseViewType;
-import uk.gov.hmcts.ccd.v2.AbstractBaseResourceTest;
+import uk.gov.hmcts.ccd.util.PathFromUrlUtil;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,7 +21,7 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.sameInstance;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-class CaseHistoryViewResourceTest extends AbstractBaseResourceTest {
+class CaseHistoryViewResourceTest {
 
     private static final String REFERENCE = "1234123412341238";
     private static final Long EVENT_ID = 100L;
@@ -67,7 +67,7 @@ class CaseHistoryViewResourceTest extends AbstractBaseResourceTest {
         final CaseHistoryViewResource resource = new CaseHistoryViewResource(caseHistoryView, REFERENCE);
 
         Optional<Link> self = resource.getLink("self");
-        assertThat(getActualPath(self.get().getHref()), equalTo(LINK_SELF));
+        assertThat(PathFromUrlUtil.getActualPath(self.get().getHref()), equalTo(LINK_SELF));
     }
 
     private CaseViewEvent newCaseViewEvent() {
