+++ src/test/java/uk/gov/hmcts/ccd/v2/internal/resource/WorkbasketInputsViewResourceTest.java
@@ -19,9 +19,9 @@ import org.springframework.hateoas.Link;
 import com.google.common.collect.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import uk.gov.hmcts.ccd.v2.AbstractBaseResourceTest;
+import uk.gov.hmcts.ccd.util.PathFromUrlUtil;
 
-class WorkbasketInputsViewResourceTest extends AbstractBaseResourceTest {
+class WorkbasketInputsViewResourceTest {
     private static final String CASE_TYPE_ID = "caseTypeId";
     private static final String LINK_SELF = String.format("/internal/case-types/%s/work-basket-inputs", CASE_TYPE_ID);
     private WorkbasketInput workbasketInput1 = aWorkbasketInput().withFieldId("field1").build();
@@ -49,7 +49,7 @@ class WorkbasketInputsViewResourceTest extends AbstractBaseResourceTest {
 
         Optional<Link> self = resource.getLink("self");
 
-        assertThat(getActualPath(self.get().getHref()), equalTo(LINK_SELF));
+        assertThat(PathFromUrlUtil.getActualPath(self.get().getHref()), equalTo(LINK_SELF));
     }
 
 }
